Добавил отдельный файл
Там написано что надо
