+++ README.md
@@ -7,4 +7,4 @@ https://docs.google.com/document/d/1PZ8qdC1cgY9etBWU31gBmUPij9MZ-olWbjWrGbJraTo/
 Задачи
 https://docs.google.com/spreadsheets/d/1ka-DjMCEskMYwHZOzJRvlgEZpGA5g-WNFh_pH3sF3y4/edit?usp=sheets_home
 
-Иконки размером `30*30` или кратно (`30*60`, `120*30`...)
+Иконки размером 30*30 или кратно (30*60, 120*30...)
